Fix RTL bidi flags
Pass the bidi flags down the stack to allow correctly rendering RTL
text.
In some cases the meaning of Paint.BIDI_DEFAULT_LTR was lost and
transformed into Paint.BIDI_LTR. That broke some cases where the text
was RTL and the flag was BIDI_DEFAULT_LTR (so it shouldn't be forced).

Bug: 37136729
Test: Existing tests pass (BIDI_DEFAULT_LTR is used from the support
lib)

Change-Id: I8e78528d87d714de57c080d623411a8afea31614