Initial work for remote layoutlib
This CL creates the API and proxies to support remote layoutlib (in a
separate process).
The classes are separated in three modules: client, common and server.

The client includes the classes meant to be used from Android Studio.
Server includes the new layoutlib server that will eventually end as a
separate jar file from the client.

TODO:
- Support custom component class loading
- Improve performance of remote streams and image transfers
- Add Android build files

Test: Adds initial test checking that it can pass renderActivity
Change-Id: I34cf050ec099282683d87b75a576e4a3ad067fdd