Fix ContextThemeWrapper
Removing the overriden implementations of obtainStyledAttributes in
Context to allow the chance of a modified Theme to be applied by the
original version. Now, internalObtainStyledAttributes will be called
from the Theme after the correct Theme has been set.
Updated the test app to be able to add a new test.

Test: New ContextThemeWrapper test in RenderTests
Change-Id: I0c05b0afe6a48c9634864814715912416c96db6a
Fixes: 66902070