Update BridgeTypedArray following platform commit 5dc45ca
Change I13f7080f introduces a new check for whether the text in a
TextView is set from the xml attribute or through a resource, by trying
to find the associated resourceId. When set from xml, no id is found,
which is the correct behaviour. So in that case, layoutlib should not
log a warning.

Test: layoutlib tests
Change-Id: Ifa7dc09ab3d7ea391711fd7fa96da6c6b83306b5