Ensure \uFEFF is given a zero width when rendering
A bug in the JDK from JetBrains used in Android Studio assigns a glyph
with non-zero width to the character \uFEFF when rendering. That is
wrong and creates rendering issues due to a bad measurement of the text
layout.
This is a workaround in layoutlib to make sure that \uFEFF is not given
a width when rendering.

Change-Id: Iab1c5080022793e034e8ccfbe7e249fcc8f35856
Fixes: 69789485
Test: manually tested within Studio
(cherry picked from commit 0fcbac3a3d933561336a02c459660cf049aa3f71)