Implement a new ImageDecoder for layoutlib
Since the constructor for ImageDecoder coming from the platform is
private, it was not possible in layoutlib to create ImageDecoder
objects, which prevented the correct handling of NinePatchDrawable
following commit 5aec67c (Change I030b4bb8).
This solves that issue by replacing the ImageDecoder class coming from
the platform by a custom one for layoutlib.

Test: layoutlib tests
Change-Id: I44599ddc1c91afb4585139272a58ec712b60e40a