Updating bars for PI preview
Status bar and navigation bar updates. Some changes on the handling of
the toolbars to allow to override the layout of a toolbar based on the
API version. Also, added support quickstep navigation toolbar.

Removed the bars from a few tests where it didn't
make sense to have them there.

Test: Updating unit tests
Change-Id: I16e2dee769d948654c04ac3f2eb58fe4ae2fce02