Do not release FontFamily reference at the builder stage
A native FontFamily object is first created as a builder before the
final object is created and the native builder freed. Layoutlib uses the
same reference for the builder and final object delegate. So the
reference should not be freed at the end of the builder stage because it
could cause the FontFamily_Delegate object to be garbage collected.

Bug: 77629275
Test: manual
Change-Id: I84686b90f56e2466a5f658155cab58d936ec2cec
(cherry picked from commit 54cff2c1941cf39d037c1d2f5e2a00503a95d5ca)