Remove sun. classes from layoutlib.jar
Layoutlib is built with JDK 9 which explains why relying on the class
loader to check for those classes is dubious, as Studio is using JDK 8
to run and there can be differences.
That is the reason why some classes from the sun package where kept in
layoutlib.jar, despite being in the JDK Studio is using, and that causes
a conflict.

Bug: 109745308
Test: N/A
Change-Id: I8f3b4b34c7297689f4fa170d1ba2487256372d5f
(cherry picked from commit 5f14d0d7bc3f04e3dd464c7f1f6821afcba98ba1)