Don't change java.text.DateFormat to android.icu.DateFormat
This was originally done because the implementation of java.text.DateFormat
and SimpleDateFormat were different between Android and the JDK in
Studio.
They are now almost identical, and also the classes from andoid.icu have
a different API as the one from java.text, which can lead to exceptions
in the layout editor rendering.

Bug: 79404948
Test: N/A
Change-Id: Iaafaa5a2fd14a4587685f6961b5d885758c4ed5f