Update layoutlib following commits 291ef05 and 55e49d8
Change Ia02c201a introduced a new parameter to a native method. We
update the corresponding delegate, though since our implementation is
far removed from the current one, we ignore the effect of this new
parameter.
Change Idd50e2f0 changed the name of a native metod, which we update,
and it added a new ColorFilter which we do not support.

Bug: 74062470
Test: layoutlib tests
Change-Id: I255d6e4b54d757e313880dbead54e2970a30900f