Register method again if class loader has changed
In PropertyValuesHolder_Delegate, do not use the cached method if the
class loader is different than the one which originally loaded it.
Instead, replace the cached method with a new one loaded with the
correct class loader.
In addition, do not clean the cache in the RenderTask tear down, as the
registered methods could be used again for AnimatedDrawables without
being first re-registered.

Bug: 130496485
Test: tested from Studio
Change-Id: Iedfbe93e3cbbc0cd52ea002249d7f0a18b6cfc1f