Do not reject any views with noop canvas
To calculate animations we initially run the rendering code with an noop
canvas. There in if block in

boolean View::draw(Canvas canvas, ViewGroup parent, long drawingTime)

that has a condition that checks wheather we can skip a widget or
proceed with the rendering. We need to proceed in order to properly
initialize animations in all the widget. Currently we proceed because of
concatMatrix variable that is incorrectly calculated from mRenderNode.
However, when the RenderNode class is properly backed by the native code
this check does not work anymore and skips the noop rendering of the
widgets (and therefore animation initialization). We have to make noop
canvas to satisfy the condition so that we proceed with the rendering
(as it happens with the proper canvas that does not reject).

Bug: N/A
Test: all tests shoudl pass
Change-Id: Iafeda6dfa076920f5503b30b3c9c011a727c5ec2
(cherry picked from commit 5f7c6c3f6acafd7f246ceb3f1fa4973682d54031)