Ensure Typeface is initialized before rendering
If Typeface is not initialized, rendering may crash if it needs a
default font. This makes sure that Typeface is initialized at the same
time as resources, so that a default font is always available.

Bug: N/A
Test: tested from Studio
Change-Id: I13cd0b4181df26c0db81808705bc49de72f87de1
(cherry picked from commit 31aa6e1016c3f92ed6a10008f6ec68db007ea3a8)