Allow font dir customization
Running tests from IntelliJ or from the test script uses two different
locations for the platform, so we cannot depend on it to get the font
location. This allows specifying the font dir location independently of
the platform dir.
This also fixes the location for the ICU data dir when running tests
from IntelliJ.

Bug: N/A
Test: layoutlib tests
Change-Id: I339f6356c55fd111c18dea922a1c0b063baa40b6