Update to accommodate variable fonts.
Author: Marek Z. Jeziorek <marekj@google.com>

Needs to be submitted before or with
https://googleplex-android-review.git.corp.google.com/c/platform/frameworks/base/+/10280447

BUGANIZER BUGS Addressed
========================
BUG: 150095308

CHANGES
=======
In a font name table is sorted by nameID, but before that, records are
sorted by platformID, etc. In a Variable Font, there are a few
namerecords (that pertain to the VF fvar axes and named instances names)
that have the legacy platformID=1 (Macintosh), which (as per OT spec)
precede the records for platformID=3 (windows unicode). The wrong
assumption that script is making is that in a "valid" font there are
always complete sets of name records for each platformID, platEncID
and langID triplets.

This is not necessarily the case.

Test: run build_font_single.py script against all possible Noto VF fonts
Test: run build_font_single.py script against all current Android fonts

Change-Id: I403a72608ad3a1034e52a51b73871e08cf6283a4
Signed-off-by: Marek Z Jeziorek <marekj@google.com>
(cherry picked from commit a327549addd4104be016ee9287e3891ed4bd81a3)