Use deferred initialization instead of native delegates
For classes that call native methods during their static initialization,
we used to delegate all their native methods to a separate class. That
required changing the behaviour of the JNI registration to link to the
delegate classes.
To avoid that behaviour and go back to using the standard JNI
registration, we instead delegate the static initialization for those
classes. That means we have to manually call the deferred static
initialization after the JNI registration.

Bug: 150115103
Test: layoutlib tests
Change-Id: If69d67608a1453fb693c83ce4d85967876108141