LayoutLib: Stop calling Bitmap#setNinePatchChunk
Bug: 150395371
Test: m layoutlib-tests

Bitmap is moving to a mainline module. Although BitmapFactory_Delegate
is in the android.graphics package, I'm not sure it's going to be a part
of the module. So remove its call to the (soon-to-be) private API
setNinePatchChunk. The constructor for Bitmap has a parameter for the
ninePatchChunk, and BitmapFactory_Delegate only calls setNinePatchChunk
once after constructing, so pass it to the constructor instead.

Change-Id: I743d154940af235339c79055b44a04ad43411357