Put atf prebuilt through layoutlib bytecode manipulation
That ensures that all the modifications done to the Android framework
classes are also done on the atf validation code, so that they stay
compatible. In particular, that means that there is no need to have a
different android.os.Build class as atf can now call the one from the
framework directly.

Bug: 154522416
Test: layoutlib and validation tests
Change-Id: I4c55a3313d600fc772bfdd92679e7e67338c71c6