[atf] refactoring API
We're thinking of introducing new APIs and a new return type:
1) API for building hierarchy only
2) Return the hierarchy only to the render result

This way we might be able to improve upon the render time inflicted by
ATF.

This CL is a refactoring to help introducing those changes later.
1) Refactor ValidatorResult creation so it's clear when hierarchy is
created and when validatorResult is created
2) Moved few files into the same package - originally we were thinking
of render-specific validation as well. No longer applicable.

Bug: 170136315
Test: All tests passed
Change-Id: I28716088445311749a25849694ca176ac45ff834