Do not send NinePatch crash to the Studio crash reporting
The crash in getOpacity can occur if the user has not correctly defined
a NinePatch. It is not useful to send this crash to the Studio crash
reporting system. It is better to bypass it, as the preview will then
fail when attempting to draw, and that error goes through the layoutlib
logging mechanism.

Bug: 142662733
Test: N/A
Change-Id: I0ec82918cf4b30efac154a92fbd4fb81835fd9ea