[atf] add pause functionality
ATF needs to be controlled separately from render params as we're
pushing for "always on" feature. Whenever it's animating atf needs to be
turned off on the fly, which cannot be controlled thru renderparams (as
some animation rendering re-uses render task and render sessions).

Adding pause feature that temporarily disables atf.

Bug: 170136315
Test: Able to test it on studio
Change-Id: Id1b2806644e25b6e47d7d77a378aba2b53a2ecf9